V2.04 in progress
This is the current code on my version of the mod. This is not available on the workshop. Probably broken. One way or another.

Cool additions:
Water Sources for Rainfall are rendered while using the water tool from Extra Landscaping tools.
Started redeveloping a custom "WaterSource" within the DrainageBasinGrid classes, but something broke at some point.